backend/requirements.txt

fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.46
pandas>=2.0.0
requests>=2.31.0
cloud-sql-python-connector[pg8000]
google-auth
openpyxl>=3.1.0
python-dotenv
pytz
phonenumbers==9.0.22